Add all settings

+++ main.py
@@ -84,41 +84,42 @@ while(True):
 
         # If Kodi want to play a video we start mpv with the video
         if 'video' in runtime.CURRENT_PATH[-1]:
-            p = subprocess.Popen(['mpv', runtime.CURRENT_PATH[-1]['video']['url']], shell=False, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-
-            output = ''
-            if config.DEPTH_EXPLORATION_MODE:
-                # We need to stop the video after some time
-                time.sleep(5)
-                p.send_signal(signal.SIGINT)
-
-                (output, err) = p.communicate()
-                retval = p.wait()  # Maybe not needed?
-
-            else:
-                # The user need to manually close mpv
-                (output, err) = p.communicate()
-                retval = p.wait()
-
-            MPV_STDOUT = output.decode('utf-8')
-            # print('MPV_STDOUT: \n' + MPV_STDOUT)
-
             # We need to go back to the last menu after the video player
             next_item = 0
 
-            if 'Exiting... (Quit)' not in MPV_STDOUT:
-                print(WARNING + ' This video does not seem to work (see log above) ' + WARNING)
-                print('')
+            if not config.DISABLE_VIDEO_PLAYER:
+                p = subprocess.Popen(['mpv', runtime.CURRENT_PATH[-1]['video']['url']], shell=False, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
 
-                runtime.ALL_REPORTED_ERROR.append({
-                    'type': 'Video',
-                    'path': current_path_pp(runtime.CURRENT_PATH),
-                    'route': bridge.TRIGGER_ERROR_ROUTE,
-                    'params': bridge.TRIGGER_ERROR_PARAMS
-                })
+                output = ''
+                if config.DEPTH_EXPLORATION_MODE:
+                    # We need to stop the video after some time
+                    time.sleep(5)
+                    p.send_signal(signal.SIGINT)
 
-                if config.EXIT_IF_ERROR:
-                    next_item = -1
+                    (output, err) = p.communicate()
+                    retval = p.wait()  # Maybe not needed?
+
+                else:
+                    # The user need to manually close mpv
+                    (output, err) = p.communicate()
+                    retval = p.wait()
+
+                MPV_STDOUT = output.decode('utf-8')
+                # print('MPV_STDOUT: \n' + MPV_STDOUT)
+
+                if 'Exiting... (Quit)' not in MPV_STDOUT:
+                    print(WARNING + ' This video does not seem to work (see log above) ' + WARNING)
+                    print('')
+
+                    runtime.ALL_REPORTED_ERROR.append({
+                        'type': 'Video',
+                        'path': current_path_pp(runtime.CURRENT_PATH),
+                        'route': bridge.TRIGGER_ERROR_ROUTE,
+                        'params': bridge.TRIGGER_ERROR_PARAMS
+                    })
+
+                    if config.EXIT_IF_ERROR:
+                        next_item = -1
 
         # We are in the basic menu case
         else:
